+++ packages/pulumi-tools/src/executors/config/schema.d.ts
@@ -1,0 +1,9 @@
+import { PulumiCommandOptions } from "../../base/base-executor";
+
+export interface ConfigExecutorSchema extends PulumiCommandOptions {
+  showSecrets?: boolean;
+  action: string;
+  secret?: boolean;
+  path?: boolean;
+  value?: string;
+}