+++ packages/pulumi-tools/src/executors/import/schema.d.ts
@@ -1,0 +1,6 @@
+import { PulumiCommandOptions } from "../../base/base-executor";
+
+export interface ImportExecutorSchema extends PulumiCommandOptions {
+  id: string;
+  target: string;
+}