+++ packages/pulumi-tools/src/executors/preview/schema.d.ts
@@ -1,0 +1,5 @@
+import { PulumiCommandOptions } from "../../base/base-executor";
+
+export interface PreviewExecutorSchema extends PulumiCommandOptions {
+  expectNoChanges?: boolean;
+}