+++ packages/pulumi-tools/src/executors/refresh/schema.d.ts
@@ -1,0 +1,6 @@
+import { PulumiCommandOptions } from "../../base/base-executor";
+
+export interface RefreshExecutorSchema extends PulumiCommandOptions {
+  skipPreview?: boolean;
+  yes?: boolean;
+}