+++ packages/pulumi-tools/src/executors/up/schema.d.ts
@@ -1,0 +1,8 @@
+import { PulumiCommandOptions } from "../../base/base-executor";
+
+export interface UpExecutorSchema extends PulumiCommandOptions {
+  skipPreview?: boolean;
+  yes?: boolean;
+  suppressOutputs?: boolean;
+  json?: boolean;
+}